% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{all_obs_types}
\alias{all_obs_types}
\title{Get a list of all observation types used in a model}
\usage{
all_obs_types(object)
}
\arguments{
\item{object}{An object of class \code{epimodel}.}
}
\description{
Get a list of all observation types used in a model
}
