% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EnglandNewCases}
\alias{EnglandNewCases}
\title{Covid-19 Case Counts for England}
\format{
A dataframe with three columns, \code{date}, \code{region} and \code{cases}. Each row gives case counts for a given date in England.
}
\usage{
EnglandNewCases
}
\description{
Contains case counts of SARS-CoV-2 in England from 30/01/2020 until 30/05/2021. Case counts correspond to
'New Cases by Specimen Date', as defined by Public Health England. The data was downloaded from
\insertCite{PHE;textual}{epidemia} on 01/06/2021. Case counts in the last few days of May may be
underreported as not all cases have been counted as of the download date.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
