% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemia-package.R
\docType{package}
\name{epidemia-package}
\alias{epidemia-package}
\alias{epidemia}
\title{Flexible Epidemic Modeling with epidemia}
\description{
The \pkg{epidemia} package allows researchers to flexibly
specify and fit Bayesian epidemiological models in the style of
\insertCite{Flaxman2020;textual}{epidemia}. The
package leverages R's formula interface to parameterize the reproduction rate
in terms of covariates, and allows pooling of parameters.
The design of the package has been inspired by, and borrowed from, the \pkg{rstanarm}
package \insertCite{goodrich_2020}{epidemia}.
\pkg{epidemia} uses \pkg{rstan} \insertCite{rstan_2020}{epidemia} as the backend for fitting the models.
The primary model fitting function in \pkg{epidemia} is \code{\link[epidemia]{epim}}.
}
\references{
\insertAllCited()
}
