% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epim.R
\name{epim}
\alias{epim}
\title{Fits an Epidemiological Model}
\usage{
epim(
  rt,
  inf,
  obs,
  data,
  algorithm = c("sampling", "meanfield", "fullrank"),
  group_subset = NULL,
  prior_PD = FALSE,
  init_run = FALSE,
  ...
)
}
\arguments{
\item{rt}{An object of class \code{\link[epidemia]{epirt}}. This specifies
the model for the time-varying reproduction number.}

\item{inf}{An object of class \code{\link[epidemia]{epiinf}}. This specifies
the model for latent infections. Defaults to \code{epiinf()}.}

\item{obs}{A list of \code{\link[epidemia]{epiobs}} objects. Each
element defines a model for the specified observation vector.}

\item{data}{A dataframe containing all data required to fit the model.
See \link{lm}.}

\item{algorithm}{One of \code{"sampling"}, \code{"meanfield"} or
\code{"fullrank"}. This specifies which \pkg{rstan} function to use for
fitting the model.}

\item{group_subset}{An optional vector specifying a subset of groups to
model. Elements should correspond to the group levels specified through the
\code{data} argument.}

\item{prior_PD}{Same as in \code{\link[rstanarm]{stan_glm}}. If \code{TRUE},
samples parameters from the prior disribution.
Defaults to \code{FALSE}.}

\item{init_run}{For certain datasets the sampler can find itself trapped in a
local mode where herd immunity is achieved. If TRUE, an MCMC run where
the population adjustment is disabled is used to initialise the parameters for
the main sampling. If TRUE, this is done with default parameters. If instead a list is
provided, these parameters are passed on to the rstan sampling function for the initial run.
initial run. The seed used is that specified in \code{init_run}, or that
specified in ..., or no seed, in that order.}

\item{...}{An (optional) named list of parameters to pass to the
\pkg{rstan} function used for model fitting, for example
\code{rstan::sampling}.}
}
\value{
An object of class "epimodel".
}
\description{
Fits a Bayesian epidemiological model specified by the \code{formula}
argument.
}
\details{
\code{epim} is the primary model fitting function in \pkg{epidemia}, and
fits models
in the style of \insertCite{Flaxman2020;textual}{epidemia}. Multiple groups
(countries/states/age cohorts)
can be modeled simultaneously using multilevel models. The time-varying
reproduction number can be paramterised by a number of covariates through
the \code{formula} argument. This is reasonably flexible. For example,
random effects terms in the style of the \pkg{lme4} package can be included.
The prior distribution on the parameters in the regression are handled
through the arguments \code{prior},
\code{prior_intercept} and \code{prior_covariance}.
}
\examples{
\dontrun{
data("EuropeCovid")

args <- EuropeCovid
args$algorithm <- "sampling"
args$rt <- epirt(
 formula=R(country, date) ~ 0 + lockdown,
 args$prior <- shifted_gamma(shape = 1 / 6,
                             scale = 1,
                             shift = log(1.05) / 6)
)

fit <- do.call("epim", args)
plot_rt(fit, group = "Germany")
}
}
\references{
\insertAllCited{}
}
