% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epimodel-objects.R
\name{epimodel-objects}
\alias{epimodel-objects}
\title{Fitted Epidemiological Model Objects}
\description{
An S3 class representing a fitted epidemiological model.
}
\details{
The workhorse function \code{epim} of the \pkg{epidemia} package returns an object
of class \code{epimodel}. This is heavily based on the \code{stanreg} class in \pkg{rstanarm}
(see \code{\link[rstanarm]{stanreg-objects}}). The internals are unimportant, but it is
helpful to read the documentation to understand how to use methods operating on \code{epimodel}
objects.
}
