% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiobs.R
\name{epiobs}
\alias{epiobs}
\title{Helper for constructing an object of class 'epiobs'}
\usage{
epiobs(
  formula,
  i2o,
  family = "neg_binom",
  link = "logit",
  center = FALSE,
  prior = rstanarm::normal(scale = 0.2),
  prior_intercept = rstanarm::normal(scale = 0.2),
  prior_aux = rstanarm::normal(location = 10, scale = 5),
  ...
)
}
\arguments{
\item{formula}{A formula defining the model for the observations.}

\item{i2o}{A probability vector with the following interpretation.
Conditional on an observation "event" (i.e. a single death or
hospitalisation etc.), the nth element represents the probability that the
individual was infected exactly n days prior to this.}

\item{family}{A string representing the error distribution for the model.
Can be "poisson", "neg_binom", "quasi_poisson", "normal" or "log_normal".}

\item{link}{A string representing the link function used to transform the
covariates. The linear predictor constructed from the covariates is
transformed by the inverse link function, then multiplied by the weighted
previous infections. This quantity represents the mean of the response
distribution.}

\item{center}{If \code{TRUE} then the covariates are centered to
have mean zero. All of the priors are then interpreted as
priors on the centered covariates. Defaults to \code{FALSE}.}

\item{prior}{Same as in \code{\link[rstanarm]{stan_glm}}. \strong{Note:}
If \code{autoscale=TRUE} in the call to the prior distribution
then automatic rescaling of the prior may take place.}

\item{prior_intercept}{Same as in \code{\link[rstanarm]{stan_glm}}. Prior
for the regression intercept, if one has been specified.}

\item{prior_aux}{Specify the prior distribution for the auxiliary parameter
if it exists. Only used if family is negative binomial (reciprocal
dispersion), quasi poisson (dispersion), normal (standard deviation) or
normal, in which case this represents the prior on the reciprocal
log normal (sigma parameter). See \code{\link[rstanarm]{stan_glm}}
for more details.}

\item{...}{Additional arguments for \code{\link[stats]{model.frame}}}
}
\description{
Defines a model for an observation vector. These observations
are taken to be a function of the latent infections in the poulation.
Examples include daily death or hospitalisation rates. For more details on
the model assumptions please refer to the online vignettes.
}
