% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epirt.R
\name{epirt}
\alias{epirt}
\title{Model Reproduction Numbers}
\usage{
epirt(
  formula,
  link = "log",
  center = FALSE,
  prior = rstanarm::normal(scale = 0.5),
  prior_intercept = rstanarm::normal(scale = 0.5),
  prior_covariance = rstanarm::decov(scale = 0.5),
  ...
)
}
\arguments{
\item{formula}{An R object of class \code{"formula"}. The left hand side
must take the form \code{R(group,date)}, with \code{group} representing a factor
vector indicating group membership (i.e. country, state, age cohort),
and \code{date} being a vector of Date objects.}

\item{link}{The link function. This must be either "identity", "log", or a call
to scaled_logit.}

\item{center}{If \code{TRUE} then the covariates for the \eqn{R_t} regression
are centered to have mean zero. All of the priors are then interpreted as
prior on the centered covariates. Defaults to \code{FALSE}.}

\item{prior}{Same as in \code{\link[rstanarm]{stan_glm}}. In addition to the
\pkg{rstanarm} provided \link[rstanarm]{priors},
a \link[epidemia]{shifted_gamma} can be used. \strong{Note:} If
\code{autoscale=TRUE} in the call to the prior distribution then
automatic rescaling of the prior may take place.}

\item{prior_intercept}{Same as in \code{\link[rstanarm]{stan_glm}}. Prior for
the regression intercept (if it exists).}

\item{prior_covariance}{Same as in \code{\link[rstanarm]{stan_glmer}}. Only
used if the \code{formula} argument specifies random effects.}

\item{...}{Additional arguments for \code{\link[stats]{model.frame}}}
}
\description{
Defines a model for the latent time varying reproduction number. For more
details on the model assumptions please refer to the online vignettes.
}
\examples{
library(epidemia)
data("EuropeCovid")
options(mc.cores = parallel::detectCores())

data <- EuropeCovid$data
data$week <- lubridate::week(data$date)

# collect arguments for epim
args <- list(
  inf = epiinf(gen = EuropeCovid$si),
  obs = epiobs(deaths ~ 1, i2o = EuropeCovid$inf2death, link = scaled_logit(0.02)),
  data = data, 
  algorithm = "fullrank", # For speed - should generally use "sampling"
  iter = 2e4,
  group_subset = "France",
  seed = 12345,
  refresh = 0
)

# a simple random walk model for R
args$rt <- epirt(
  formula = R(country, date) ~ rw(time = week),
  link = scaled_logit(7)
)

fm1 <- do.call(epim, args)
plot_rt(fm1)

# Modeling effects of NPIs
args$rt <- epirt(
  formula = R(country, date) ~ 1 + lockdown + public_events,
  link = scaled_logit(7)
)

fm2 <- do.call(epim, args)
plot_rt(fm2)


# shifted gamma prior for NPI effects
args$rt <- epirt(
  formula = R(country, date) ~ 1 + lockdown + public_events,
  link = scaled_logit(7),
  prior = shifted_gamma(shape = 1/2, scale = 1, shift = log(1.05)/2)
)

# How does the implied prior look?
args$prior_PD <- TRUE
fm3 <- do.call(epim, args)
plot_rt(fm3)
}
