% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EuropeCovid2}
\alias{EuropeCovid2}
\title{Covid-19 data for European countries}
\format{
A named list. The fields are:
\describe{
\item{data}{A data frame giving indicators of certain non-pharmaceutical interventions in each country, along with death data and populations.}
\item{inf2death}{A numeric vector representing the time distribution from infection to death.}
\item{si}{The serial interval of covid-19 assumed in \insertCite{Flaxman2020;textual}{epidemia}.}
}
}
\usage{
EuropeCovid2
}
\description{
Similar to \code{EuropeCovid}, with the following exceptions. Daily death data is obtained from the WHO COVID-19 Explorer as of 05/01/2021. This differs
from the data used in \insertCite{Flaxman2020;textual}{epidemia}, because counts were updated retrospectively by the WHO as new information came
to light. Daily case data is also included from the same source. This data runs from 03/01/2020 until 30/06/2020.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
