% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{evaluate_forecast}
\alias{evaluate_forecast}
\title{Posterior model evaluations}
\usage{
evaluate_forecast(
  object,
  newdata = NULL,
  type,
  groups = NULL,
  metrics = NULL,
  levels = c(50, 95)
)
}
\arguments{
\item{object}{A fitted model object returned by \code{epim}.
See \code{\link{epimodel-objects}}.}

\item{newdata}{If provided, the original \code{data} used
in \code{object} is overidden. Useful for forecasting}

\item{type}{the name of the observations to plot. This should match one
of the names of the \code{obs} argument to \code{epim}.}

\item{groups}{\code{NULL}, string or character vector specifying which groups
to plot. Default is \code{NULL}, which plots all possible groups.}

\item{metrics}{A string or character vector specifying the plotted
forecast error metrics. One of \code{NULL}, \code{"crps"},
\code{"mean_abs_error"}}

\item{levels}{numeric vector giving the levels of the plotted
credible intervals}
}
\value{
A named list with dataframes giving metrics and coverage.
}
\description{
Calculate daily error using one of three metrics, and also return coverage
of credible intervals. Uses continuous ranked probability
score (CRPS), mean absolute error and median absolute error.
}
