% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epimodel-methods.R
\name{get_x}
\alias{get_x}
\alias{get_z}
\title{Extract X or Z from an epimodel object}
\usage{
get_x(object, ...)

get_z(object, ...)
}
\arguments{
\item{object}{A fitted model object returned by \code{\link{epim}}.
See \code{\link{epimodel-objects}}.}

\item{...}{Other arguments passed to methods.}
}
\value{
A matrix.
}
\description{
Extract X or Z from an epimodel object
}
