% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epimodel-methods.R
\name{ngrps.mixed}
\alias{ngrps.mixed}
\title{Returns the levels for each grouping factor in the fitted object}
\usage{
\method{ngrps}{mixed}(object, ...)
}
\arguments{
\item{object}{an \code{R} object, see Details.}

\item{...}{currently ignored.}
}
\value{
The number of levels (of a factor) or vector of number of levels for
  each \dQuote{grouping factor} of a
}
\description{
Returns the levels for each grouping factor in the fitted object
}
