% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Plot coverage probability of posterior credible intervals}
\usage{
plot_coverage(
  object,
  type,
  newdata = NULL,
  groups = NULL,
  levels = c(50, 95),
  period = NULL,
  by_group = FALSE,
  by_unseen = FALSE,
  plotly = FALSE
)
}
\arguments{
\item{object}{A fitted model object returned by \code{epim}.
See \code{\link{epimodel-objects}}.}

\item{type}{the name of the observations to plot. This should match one
of the names of the \code{obs} argument to \code{epim}.}

\item{newdata}{If provided, the original \code{data} used
in \code{object} is overidden. Useful for forecasting}

\item{groups}{\code{NULL}, string or character vector specifying which groups
to plot. Default is \code{NULL}, which plots all possible groups.}

\item{levels}{numeric vector giving the levels of the plotted
credible intervals}

\item{period}{Buckets computed empirical probabilities into time periods
if specified.}

\item{by_group}{Plot coverage for each group individually}

\item{by_unseen}{Plot coverage separately for seen and unseen observations.
Observations are 'seen' if they were used for fitting.}

\item{plotly}{If TRUE, wraps the ggplot object into a plotly object, for
interactive graphing.}
}
\description{
Plots histograms showing empirical coverage of credible intervals
specified using 'levels'. Can bucket by time period, by group, by
whether the observation is new (not used in fitting).
}
