% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_epi.R
\name{plot_infections}
\alias{plot_infections}
\alias{plot_infections.epimodel}
\title{Plotting the underlying number of infections over time}
\usage{
plot_infections(object, ...)

\method{plot_infections}{epimodel}(
  object,
  groups = NULL,
  dates = NULL,
  date_breaks = "2 weeks",
  date_format = "\%Y-\%m-\%d",
  cumulative = FALSE,
  levels = c(30, 60, 90),
  log = FALSE,
  plotly = FALSE,
  ...
)
}
\arguments{
\item{object}{A fitted model object returned by \code{epim}.
See \code{\link{epimodel-objects}}.}

\item{...}{Additional arguments for \code{\link[epidemia]{posterior_infections}}. Examples include \code{newdata}, which allows
predictions or counterfactuals.}

\item{groups}{\code{NULL}, string or character vector specifying which groups
to plot. Default is \code{NULL}, which plots all possible groups.}

\item{dates}{a vector of (start date, end date) defining the date range
to be plotted. Must be coercible to date if not NA. If NA, this means
use the lower/upper limit as appropriate. See examples.}

\item{date_breaks}{string giving the distance between date tick labels.
Default is "2 weeks".
Passed as \code{date_breaks} argument to \code{ggplot::scale_x_date} -
see \url{https://ggplot2.tidyverse.org/reference/scale_date.html}}

\item{date_format}{the date format for coercing \code{dates}.
Default is "\%Y-\%m-\%d"}

\item{cumulative}{If TRUE, plots the cumulative observations.}

\item{levels}{numeric vector giving the levels of the plotted
credible intervals}

\item{log}{If TRUE, plots the observations on a pseudo-linear scale.}

\item{plotly}{If TRUE, wraps the ggplot object into a plotly object, for
interactive graphing.}
}
\value{
A ggplot object which can be further modified.
}
\description{
Plots credible intervals for the underlying number of infections.
The user can control the levels of the intervals and the plotted group(s).
This is a generic function.
}
\examples{
\dontrun{
## load required data
library(epidemia)
data("EuropeCovid")
## setup sampling
args <- EuropeCovid
args$algorithm <- "sampling"
args$sampling_args <- list(iter=1e3,seed=12345)
args$group_subset <- c("Italy", "Austria", "Germany")
args$rt <- epirt(
  formula = R(country, date) ~ 1 + lockdown,
  prior = rstanarm::normal(location=0, scale=.5),
  prior_intercept = rstanarm::normal(location=0, scale=2) 
)

## run sampling
fit <- do.call("epim", args)

## make plots
plot_infections(fit) # default, plots all groups and dates
plot_infections(fit, 
                dates=c("2020-03-21", NA)) # plot 21 March 2020 onwards
plot_infections(fit, 
                dates=c(NA, "2020-03-20")) # plot up to  20 March 2020
plot_infections(fit, 
                dates=c("2020-03-20", "2020-04-20")) # plot 20 March-20 April 2020
plot_infections(fit, 
                dates=c("2020-03-20", "2020-04-20"), 
                date_breaks="1 day") # plot 21 March-20 April 2020 with ticks every day
plot_infections(fit, 
                dates=c("2020-03-20", "2020-04-20"),
                date_breaks="1 week") # plot 21 March-20 April 2020 with ticks every week
plot_infections(fit, 
                dates=c("2020-20-03", "2020-20-04"), 
                date_format="\%Y-\%d-\%m") # plot 21 March-20 April 2020 (different date format)
}
}
