% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_epi.R
\name{plot_infectious}
\alias{plot_infectious}
\alias{plot_infectious.epimodel}
\title{Plotting the underlying total infectiousness over time}
\usage{
plot_infectious(object, ...)

\method{plot_infectious}{epimodel}(
  object,
  groups = NULL,
  dates = NULL,
  date_breaks = "2 weeks",
  date_format = "\%Y-\%m-\%d",
  levels = c(30, 60, 90),
  log = FALSE,
  plotly = FALSE,
  ...
)
}
\arguments{
\item{object}{A fitted model object returned by \code{epim}.
See \code{\link{epimodel-objects}}.}

\item{...}{Additional arguments for
\code{\link[epidemia]{posterior_infectious}}. Examples include
\code{newdata}, which allows predictions or counterfactuals.}

\item{groups}{\code{NULL}, string or character vector specifying which groups
to plot. Default is \code{NULL}, which plots all possible groups.}

\item{dates}{a vector of (start date, end date) defining the date range
to be plotted. Must be coercible to date if not NA. If NA, this means
use the lower/upper limit as appropriate. See examples.}

\item{date_breaks}{string giving the distance between date tick labels.
Default is "2 weeks".
Passed as \code{date_breaks} argument to \code{ggplot::scale_x_date} -
see \url{https://ggplot2.tidyverse.org/reference/scale_date.html}}

\item{date_format}{the date format for coercing \code{dates}.
Default is "\%Y-\%m-\%d"}

\item{levels}{numeric vector giving the levels of the plotted
credible intervals}

\item{log}{If TRUE, plots the observations on a pseudo-linear scale.}

\item{plotly}{If TRUE, wraps the ggplot object into a plotly object, for
interactive graphing.}
}
\value{
A ggplot object which can be further modified.
}
\description{
Plots credible intervals for the total infectiousness over time. This is
defined as the sum of each infected person, weighted by how infectious each
individual is, given how long they have been infected for.The user can
control the levels of the intervals and the plotted group(s).
This is a generic function.
}
