% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{plot_metrics}
\alias{plot_metrics}
\title{Plot CRPS, Median/Mean Absolute Error}
\usage{
plot_metrics(
  object,
  groups = NULL,
  type,
  metrics = NULL,
  newdata = NULL,
  plotly = FALSE
)
}
\arguments{
\item{object}{A fitted model object returned by \code{epim}.
See \code{\link{epimodel-objects}}.}

\item{groups}{\code{NULL}, string or character vector specifying which groups
to plot. Default is \code{NULL}, which plots all possible groups.}

\item{type}{the name of the observations to plot. This should match one
of the names of the \code{obs} argument to \code{epim}.}

\item{metrics}{A string or character vector specifying the plotted
forecast error metrics. One of \code{NULL}, \code{"crps"},
\code{"mean_abs_error"}}

\item{newdata}{If provided, the original \code{data} used
in \code{object} is overidden. Useful for forecasting}

\item{plotly}{Convert ggplot into plotly}
}
\value{
A ggplot object which can be further modified.
}
\description{
Plots various metrics for evaluating probabilistic forecasts by group.
}
