% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_epi.R
\name{plot_obs}
\alias{plot_obs}
\alias{plot_obs.epimodel}
\title{Plotting the posterior predictive distribution}
\usage{
plot_obs(object, ...)

\method{plot_obs}{epimodel}(
  object,
  type,
  posterior_mean = FALSE,
  groups = NULL,
  dates = NULL,
  date_breaks = "2 weeks",
  date_format = "\%Y-\%m-\%d",
  cumulative = FALSE,
  bar = TRUE,
  levels = c(30, 60, 90),
  log = FALSE,
  plotly = FALSE,
  ...
)
}
\arguments{
\item{object}{A fitted model object returned by \code{epim}.
See \code{\link{epimodel-objects}}.}

\item{...}{Additional arguments for
\code{\link[epidemia]{posterior_predict.epimodel}}. Examples include
\code{newdata}, which allows
predictions or counterfactuals.}

\item{type}{the name of the observations to plot. This should match one
of the names of the \code{obs} argument to \code{epim}.}

\item{posterior_mean}{If true, the credible intervals are plotted for the
posterior mean. Defaults to FALSE,
in which case the posterior predictive is plotted.}

\item{groups}{\code{NULL}, string or character vector specifying which groups
to plot. Default is \code{NULL}, which plots all possible groups.}

\item{dates}{a vector of (start date, end date) defining the date range
to be plotted. Must be coercible to date if not NA. If NA, this means
use the lower/upper limit as appropriate. See examples.}

\item{date_breaks}{string giving the distance between date tick labels.
Default is "2 weeks".
Passed as \code{date_breaks} argument to \code{ggplot::scale_x_date} -
see \url{https://ggplot2.tidyverse.org/reference/scale_date.html}}

\item{date_format}{the date format for coercing \code{dates}.
Default is "\%Y-\%m-\%d"}

\item{cumulative}{If TRUE, plots the cumulative observations.}

\item{bar}{If TRUE, observations are plotted as a bar plot. Otherwise,
a scatterplot is used.}

\item{levels}{numeric vector giving the levels of the plotted
credible intervals}

\item{log}{If TRUE, plots the observations on a pseudo-linear scale.}

\item{plotly}{If TRUE, wraps the ggplot object into a plotly object, for
interactive graphing.}
}
\value{
A ggplot object which can be further modified.
}
\description{
Plots credible intervals for the observed data under the posterior
predictive distribution, and for a specific observation type.
The user can control the levels of the intervals and the plotted group(s).
This is a generic function.
}
\examples{
\dontrun{
## load required data
library(epidemia)
data("EuropeCovid")
## setup sampling
args <- EuropeCovid
args$algorithm <- "sampling"
args$sampling_args <- list(iter=1e3,seed=12345)
args$group_subset <- c("Italy", "Austria", "Germany")
args$rt <- epirt(
  formula = R(country, date) ~ 1 + lockdown,
  prior = rstanarm::normal(location=0, scale=.5),
  prior_intercept = rstanarm::normal(location=0, scale=2) 
)

## run sampling
fit <- do.call("epim", args)

## make plots
plot_obs(fit, type="deaths")
plot_obs(fit, type="deaths",
              dates=c("2020-03-21", NA))
plot_obs(fit,
         type="deaths",
         dates=c(NA, "2020-03-20"))
plot_obs(fit,
         type="deaths",
         dates=c("2020-03-20", "2020-04-20"))
plot_obs(fit,
         type="deaths",
         dates=c("2020-03-20", "2020-04-20"),
         date_breaks="1 day")
plot_obs(fit,
         type="deaths",
         dates=c("2020-03-20", "2020-04-20"),
         date_breaks="1 week")
plot_obs(fit,
         type="deaths",
         dates=c("2020-20-03", "2020-20-04"),
         date_format="\%Y-\%d-\%m")
}
}
