% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_epi.R
\name{plot_rt}
\alias{plot_rt}
\alias{plot_rt.epimodel}
\title{Plotting the time-varying reproduction rates}
\usage{
plot_rt(object, ...)

\method{plot_rt}{epimodel}(
  object,
  groups = NULL,
  step = FALSE,
  dates = NULL,
  date_breaks = "2 weeks",
  date_format = "\%Y-\%m-\%d",
  levels = c(30, 60, 90),
  log = FALSE,
  smooth = 1,
  plotly = FALSE,
  ...
)
}
\arguments{
\item{object}{A fitted model object returned by \code{epim}.
See \code{\link{epimodel-objects}}.}

\item{...}{Additional arguments for \code{\link[epidemia]{posterior_rt}}.
Examples include \code{newdata}, which allows predictions or
counterfactuals. \code{adjusted=FALSE} prevents application of
the population adjustment to the reproduction number.}

\item{groups}{\code{NULL}, string or character vector specifying which groups
to plot. Default is \code{NULL}, which plots all possible groups.}

\item{step}{If true, draw median and CIs as a step function.}

\item{dates}{a vector of (start date, end date) defining the date range
to be plotted. Must be coercible to date if not NA. If NA, this means
use the lower/upper limit as appropriate. See examples.}

\item{date_breaks}{string giving the distance between date tick labels.
Default is "2 weeks".
Passed as \code{date_breaks} argument to \code{ggplot::scale_x_date} -
see \url{https://ggplot2.tidyverse.org/reference/scale_date.html}}

\item{date_format}{the date format for coercing \code{dates}.
Default is "\%Y-\%m-\%d"}

\item{levels}{numeric vector giving the levels of the plotted
credible intervals}

\item{log}{whether to plot the reproduction number on a log10-scale.
Logical, default is \code{FALSE}.}

\item{smooth}{integer specifying the window used to smooth the Rt values.
Default is 1 (no smoothing).}

\item{plotly}{If TRUE, wraps the ggplot object into a plotly object, for
interactive graphing.}
}
\value{
A ggplot object which can be further modified.
}
\description{
Plots credible intervals for the time-varying reproduction rate.
The user can control the levels of the intervals and the plotted group(s).
This is a generic function.
}
\examples{
\dontrun{
## load required data
library(epidemia)
data("EuropeCovid")
## setup sampling
args <- EuropeCovid
args$algorithm <- "sampling"
args$sampling_args <- list(iter=1e3,seed=12345)
args$group_subset <- c("Italy", "Austria", "Germany")
args$rt <- epirt(
  formula = R(country, date) ~ 1 + lockdown,
  prior = rstanarm::normal(location=0, scale=.5),
  prior_intercept = rstanarm::normal(location=0, scale=2) 
)

## run sampling
fit <- do.call("epim", args)

## make plots
plot_rt(fit) # default, plots all groups and dates
plot_rt(fit, dates=c("2020-03-21", NA)) # plot 21 March 2020 onwards
plot_rt(fit, dates=c(NA, "2020-03-20")) # plot up to  20 March 2020
plot_rt(fit, dates=c("2020-03-20", "2020-04-20"))
plot_rt(fit,
        dates=c("2020-03-20", "2020-04-20"),
        date_breaks="1 day") # ticks every day
plot_rt(fit,
        dates=c("2020-03-20", "2020-04-20"),
        date_breaks="1 week") # ticks every week
plot_rt(fit,
        dates=c("2020-20-03", "2020-20-04"),
        date_format="\%Y-\%d-\%m") # (different date format)
}
}
