% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_latents.R
\name{posterior_infectious}
\alias{posterior_infectious}
\alias{posterior_infectious.epimodel}
\title{Generic function for getting posterior draws of total infectiousness
over time}
\usage{
posterior_infectious(object, ...)

\method{posterior_infectious}{epimodel}(object, newdata = NULL, draws = NULL, seed = NULL, ...)
}
\arguments{
\item{object}{A fitted model object returned by \code{\link{epim}}.
See \code{\link{epimodel-objects}}.}

\item{...}{Not used.}

\item{newdata}{If provided, the original \code{data} used
in \code{object} is overidden. Useful both for conterfactual
and prediction analysis}

\item{draws}{Number of posterior draws to use. Defaults to the number
of parameter draws in the fitted model.}

\item{seed}{An optional seed.}
}
\value{
A named list with draws of the series, and vectors giving the
corresponding dates and population.
}
\description{
Generic function for getting posterior draws of total infectiousness
over time
}
