% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_latents.R
\name{posterior_rt}
\alias{posterior_rt}
\alias{posterior_rt.epimodel}
\title{Generic function for getting posterior draws of the time-varying
reproduction rates}
\usage{
posterior_rt(object, ...)

\method{posterior_rt}{epimodel}(
  object,
  newdata = NULL,
  draws = NULL,
  seed = NULL,
  adjusted = TRUE,
  ...
)
}
\arguments{
\item{object}{A fitted model object returned by \code{\link{epim}}.
See \code{\link{epimodel-objects}}.}

\item{...}{Not used.}

\item{newdata}{If provided, the original \code{data} used
in \code{object} is overridden. Useful both for conterfactual
and prediction analysis}

\item{draws}{Number of posterior draws to use. Defaults to the number
of parameter draws in the fitted model.}

\item{seed}{An optional seed.}

\item{adjusted}{Flag whether to return the adjusted reproduction rates.
Defaults to TRUE.}
}
\value{
A named list with draws of the series, and vectors giving the
corresponding dates and population.
}
\description{
Generic function for getting posterior draws of the time-varying
reproduction rates
}
