% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.epimodel}
\alias{print.epimodel}
\title{Print fitted model details}
\usage{
\method{print}{epimodel}(x, digits = 1, ...)
}
\arguments{
\item{x}{A fitted model object returned by \code{\link{epim}}.
See \code{\link{epimodel-objects}}.}

\item{digits}{Number of decimal places to print.}

\item{...}{Not used.}
}
\value{
No return value.
}
\description{
Prints estimated regression parameters, and other model parameters.
Similar to printing of \code{rstan::stanreg} objects.
}
