% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_summary.R
\name{print.prior_summary.epimodel}
\alias{print.prior_summary.epimodel}
\title{Print method for \code{prior_summary.epimodel} objects}
\usage{
\method{print}{prior_summary.epimodel}(x, digits, ...)
}
\arguments{
\item{x}{A fitted model object returned by \code{\link{epim}}.
See \code{\link{epimodel-objects}}.}

\item{digits}{Number of decimal places to print.}

\item{...}{Not used.}
}
\value{
No return value.
}
\description{
Print method for \code{prior_summary.epimodel} objects
}
