% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_summary.R
\name{prior_summary.epimodel}
\alias{prior_summary.epimodel}
\title{Returns a summary of the prior distributions used}
\usage{
\method{prior_summary}{epimodel}(object, digits = 3, ...)
}
\arguments{
\item{object}{The object to use.}

\item{digits}{Number of digits used for rounding.}

\item{...}{Arguments passed to methods. See the methods in the \pkg{rstanarm}
package for examples.}
}
\value{
\code{prior_summary()} methods should return an object containing
information about the prior distribution(s) used for the given model.
The structure of this object will depend on the method.

The default method just returns \code{object$prior.info}, which is
\code{NULL} if there is no \code{'prior.info'} element.
}
\description{
Returns a summary of the prior distributions used
}
