% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_state.R
\name{regional_state}
\alias{regional_state}
\title{Regional State Model}
\usage{
regional_state(stan_data, ...)
}
\arguments{
\item{stan_data}{input data required for sampling}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Regional State Model
}
