% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epimodel-methods.R
\name{terms.epimodel}
\alias{terms.epimodel}
\title{Terms method for epimodel objects}
\usage{
\method{terms}{epimodel}(x, fixed.only = TRUE, random.only = FALSE, ...)
}
\arguments{
\item{x, fixed.only, random.only, ...}{See \code{\link[lme4]{terms.merMod}}}
}
\description{
Terms method for epimodel objects
}
