% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiinf.R
\name{epiinf}
\alias{epiinf}
\title{Modeling Latent Infections}
\usage{
epiinf(
  gen,
  seed_days = 6L,
  latent = FALSE,
  family = "log-normal",
  prior_aux = rstanarm::normal(10, 5),
  prior_tau = rstanarm::exponential(0.03),
  pop_adjust = FALSE,
  susceptibles = NULL
)
}
\arguments{
\item{gen}{A vector representing the generation distribution of the disease (a
probability vector).}

\item{seed_days}{Number of days for which to seed infections.}

\item{latent}{If \code{TRUE}, infections are treated as unknown parameters
to be sampled. If \code{FALSE}, the model only consider their expected value given
the reproduction numbers and seeded infections. Defaults to \code{FALSE}.}

\item{family}{Specifies prior family for latent infections.
Only used if \code{latent = TRUE}, and is currently restricted to be "log-normal".}

\item{prior_aux}{Prior distribution for the coefficient of dispersion \eqn{d} for
the offspring distribution. Only used if \code{latent = TRUE}. The number of
offspring of a given infection is assumed to have mean \eqn{\mu} and variance \eqn{d \mu}.
This argument specifies prior on \eqn{d}. Higher values of \eqn{d} imply more
super-spreading events.}

\item{prior_tau}{The prior for \eqn{\tau}.This parameter is described in the
introductory vignette, and controls the variability in the number of
seeded infections at  the beginning of the epidemic. Must be a call to
\code{\link[rstanarm]{exponential}}.}

\item{pop_adjust}{The population adjustment is a major contributor to algorithm
runtime. Although it should be implemented for a final model run, it may be
quicker to develop models without the adjustment. Defaults to TRUE.}

\item{susceptibles}{A character vector giving the name of the column in the dataframe
passed as the data argument of epim, that corresponds to the susceptible population over time.
Only used if pop_adjust=TRUE.}
}
\description{
Latent infections can be given a full prior distribution and inferred.
\code{epiinf} allows you to define this prior distribution. The user can also
specify models where only the expected infections are used, avoiding the need
to perform full Bayesian inference for these quantities.
}
