% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epimodel-methods.R
\name{formula.epimodel}
\alias{formula.epimodel}
\title{formula method for epimodel objects}
\usage{
\method{formula}{epimodel}(x, ...)
}
\arguments{
\item{x}{An epimodel object.}

\item{...}{Can contain \code{fixed.only} and \code{random.only} arguments
that both default to \code{FALSE}.}
}
\description{
formula method for epimodel objects
}
