% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epimodel-methods.R
\name{model.frame.epimodel}
\alias{model.frame.epimodel}
\title{model.frame method for epimodel objects. Please see  \code{\link[stats]{model.frame}}
for more details.}
\usage{
\method{model.frame}{epimodel}(formula, fixed.only = FALSE, ...)
}
\arguments{
\item{formula}{A fitted model object returned by \code{epim}.
See \code{\link{epimodel-objects}}.}

\item{fixed.only}{See \code{\link[lme4]{model.frame.merMod}}.}

\item{...}{See \code{\link[stats]{model.frame}}.}
}
\description{
model.frame method for epimodel objects. Please see  \code{\link[stats]{model.frame}}
for more details.
}
