% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{posterior_sample_size}
\alias{posterior_sample_size}
\title{Get posterior sample size from a fitted model}
\usage{
posterior_sample_size(object)
}
\arguments{
\item{object}{An object of class \code{epimodel}}
}
\description{
Get posterior sample size from a fitted model
}
