% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemia-package.R, R/plots_epi.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ngrps}
\alias{posterior_predict}
\alias{prior_summary}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}

  \item{lme4}{\code{\link[lme4]{ngrps}}}

  \item{rstantools}{\code{\link[rstantools]{posterior_predict}}, \code{\link[rstantools]{prior_summary}}}
}}

